refactor(ui): split profileContext.tsx (914 → 4 files) by concern
Twelfth worst-first file-size refactor. The ProfileProvider's ~30 useCallback
hooks split into three focused hook modules, leaving profileContext.tsx as a
slim provider that just composes them.

  profileContext.tsx               292  context type + ProfileProvider assembly + public hooks
  useProfileApi.ts                 194  CRUD / import / export wrappers around React Query
  useProfileSettingsUpdates.ts     164  per-section auto-save updaters
  useProfileInterfaces.ts          310  ethernet / wifi multi-interface helpers

All under the 800-line TSX warn budget. No behaviour change: the public
ProfileContextValue surface is unchanged, callers of useProfileContext keep
working as-is.

Vite build clean; biome check clean.
